Don't inject CSS directly into the head; added CSS file for real this time
Before we were injecting CSS directly into the head of *every* admin page. Now, we look to see if it's a document page, and if so
conditionally, inject body tags into the admin body class. Every page will have the .document body class, and if a post is locked, it will have the
.document-locked body class.