Only cache revisions output if revisions exist
In some cases, and likely due to some other customization by
another plugin, it's possible to generate a post save without
entering a title or content and with no revision.

It appears that in these cases, the `get_revision_indices()`
method attempts to cache an `$output` variable that doesn't exist.
Shortly after, the `get_revision_number()` method attempts to
`array_search()` on a piece of data that is not an array.

This changes sets `$output` to an empty array and, if no data is
stored in it, avoids storing it in cache and returns it as is.