Be more explicit where possible in `verify_post_type()`
The `verify_post_type()` method assumes `false` if no other data
is passed to it and attempts to determine a valid post ID based on
the global data available to it. In some cases, this determination
can be wrong due to how things are passed around in WordPress.

In cases where a full `WP_Post` object is available or where the
post ID is provided, we can be a bit more explicit to skip the
checking of global data.

In some cases, we really do need to pay attention to the global
scope, those are left as is.