Double check the global post lookup before continuing
If we're looking to global data to find a document in the right
context, then we should not continue if the global data still
comes back empty. Previously, the results from `get_post_type()`
with empty data would not be reliable and could effect other parts
of the page load in strange ways.