Provide a way to filter or skip mime type detection
When using a plugin such as S3 Uploads, the AWS S3 SDK stream
wrapper for PHP does not include support for `mime_content_type`
and instead has its own method available for mime type detection.

Additionally, in some server configurations it's possible that
the `Content-Type` header set here does not make its way to the
response and accounts instead for extra overhead.

To allow support for situations like this, it should be possible
to filter the value of `$mimetype` before it is attempted by WP
Document Revisions.

When filtered to `false`, we skip the mime type detection all
together. When filtered to a string, the content type header is
sent.