This is to address issue #116 and possibly #82 the 404 obtained when using a permalink.
When using a permalink, the process serve_file uses get_attached_file which uses a cached version of the upload directory without direct reference to the specific post.
This gives the standard directory. If the user has defined a different directory then this is not used  - and the 404 results.

A solution would be to invoke a filter on get_attached_file - but as it will essentially have to do the str_replace on the file name, the code is put directly