Support both Edit Flow and PublishPress
PublishPress is a clone of EditFlow so this change is relatively simple.

I have taken the opportunity to use the hooks provided by these plugins to ensure the review process for these plugins only occurs if the plugins are installed and active.

If appropriate the removal process takes place immediately so don't need to worry where to hang that process.